Changed the way QuickAdd widget is rendered
Django uses nice templates to render widgets https://docs.djangoproject.com/en/1.11/ref/forms/widgets/#select
